Moved traces.loader to traces.waveform.
