+++ phy/traces/waveform.py
